+++ stroom-cluster/stroom-cluster-lock-impl-db/src/main/resources/stroom/cluster/lock/impl/db/migration/V07_00_00_01__cluster_lock.sql
@@ -1,3 +1,6 @@
+-- Stop NOTE level warnings about objects (not)? existing
+SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0;
+
 --
 -- Create the cluster_lock table
 --
@@ -7,3 +10,5 @@ CREATE TABLE IF NOT EXISTS cluster_lock (
   name                  varchar(255) NOT NULL,
   PRIMARY KEY           (id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+
+SET SQL_NOTES=@SQL_NOTES;