+++ stroom-meta/stroom-meta-impl-db/src/main/resources/stroom/meta/impl/db/migration/V07_00_00_04__meta.sql
@@ -1,3 +1,6 @@
+-- Stop NOTE level warnings about objects (not)? existing
+SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0;
+
 --
 -- Create the meta table
 --
@@ -44,3 +47,5 @@ END//
 DELIMITER ;
 CALL copy_meta();
 DROP PROCEDURE copy_meta;
+
+SET SQL_NOTES=@SQL_NOTES;